api2: refactoring
